Separate info points in new folder
