+++ pages/index.tsx
@@ -34,7 +34,7 @@ import BigNumber from 'bignumber.js';
 import { getCategories } from 'modules/polling/helpers/getCategories';
 import { InternalLink } from 'modules/app/components/InternalLink';
 import MeetDelegates from 'modules/delegates/components/MeetDelegates';
-import InformationParticipateMakerGovernance from 'modules/home/components/InformationParticipateMakerGovernance';
+import InformationParticipateMakerGovernance from 'modules/home/components/InformationParticipateMakerGovernance/InformationParticipateMakerGovernance';
 import { DelegateStatusEnum } from 'modules/delegates/delegates.constants';
 import { useBreakpointIndex } from '@theme-ui/match-media';
 import { useMkrOnHat } from 'modules/executive/hooks/useMkrOnHat';
